rename NotaryV2 to Notary
